% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries}
\alias{birdrecoveries}
\title{Dataset birdrecoveries}
\format{A data frame [3795 x 30]
\describe{
  \item{name}{}
  \item{sciname}{}
  \item{ringing_sex_code}{}
  \item{ringing_sex}{}
  \item{ringing_age_code}{}
  \item{ringing_age}{}
  \item{ringing_date}{}
  \item{ringing_lat}{}
  \item{ringing_lon}{}
  \item{ringing_country}{}
  \item{ringing_province}{}
  \item{ringing_majorregion}{}
  \item{ringing_minorregion}{}
  \item{ringing_accu2}{}
  \item{recovery_date}{}
  \item{recovery_accu_code}{}
  \item{recovery_accu}{}
  \item{recovery_lat}{}
  \item{recovery_lon}{}
  \item{recovery_accu2}{}
  \item{recovery_country}{}
  \item{recovery_province}{}
  \item{recovery_majorregion}{}
  \item{recovery_minorregion}{}
  \item{distance}{}
  \item{direction}{}
  \item{days}{}
  \item{hours}{}
  \item{recovery_code}{}
  \item{recovery_det}{}
  ...
  }}
\source{
\url{http://}
}
\usage{
birdrecoveries
}
\description{
This dataset contains all fields (both english and swedish versions)
Date: Sun Nov 13 13:38:49 2016
}
\keyword{datasets}

