% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries_eng}
\alias{birdrecoveries_eng}
\title{Dataset birdrecoveries_eng}
\format{
A data frame
}
\source{
\url{http://fagel3.nrm.se/fagel/aterfynd/SQLDataExport.csv}
}
\description{
A dataset containing bird recoveries in English.
}
\keyword{datasets}
