% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries_i18n}
\alias{birdrecoveries_i18n}
\title{Dataset birdrecoveries_i18n}
\format{
A data frame
}
\description{
A dataset containing translations for the column names.
}
\keyword{datasets}
