% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{i18n}
\alias{i18n}
\title{Translate strings for display in user interface}
\usage{
i18n(field, lang = "eng")
}
\arguments{
\item{field}{the name of the field in birdrecoveries_i18n,
defined in data-raw/translation.csv originally}

\item{lang}{the language iso3 string in lower caps, default "eng"}
}
\value{
a vector of field names if called without arguments,
otherwise the string corresponding to the field name for the
specified language
}
\description{
Translate strings for display in user interface
}
\examples{
 fields <- i18n()
}
