% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_falsterbo}
\alias{scrape_checklist_falsterbo}
\title{Scrape checklist from Falsterbo}
\usage{
scrape_checklist_falsterbo()
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
