% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_norway}
\alias{scrape_checklist_norway}
\title{Scrape checklist from Norway}
\usage{
scrape_checklist_norway()
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
