% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_ottenby}
\alias{scrape_checklist_ottenby}
\title{Scrape checklist from Ottenby}
\usage{
scrape_checklist_ottenby()
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
