% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_recoveries_falsterbo}
\alias{scrape_recoveries_falsterbo}
\title{Scrape recoveries from Falsterbo}
\usage{
scrape_recoveries_falsterbo(species = "ACNIS")
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
