% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_recoveries_norway}
\alias{scrape_recoveries_norway}
\title{Scrape recoveries from Norway}
\usage{
scrape_recoveries_norway(species_id = 1580)
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
