% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_recoveries_ottenby}
\alias{scrape_recoveries_ottenby}
\title{Scrape recoveries from Ottenby}
\usage{
scrape_recoveries_ottenby()
}
\description{
This function is deprecated because the website is no longer available.
}
\keyword{internal}
