% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swedishbirdrecoveries-package.R
\docType{package}
\name{swedishbirdrecoveries-package}
\alias{swedishbirdrecoveries-package}
\alias{swedishbirdrecoveries}
\title{swedishbirdrecoveries - An R package with Swedish Bird Recaptures}
\description{
The swedishbirdrecoveries package provides tools for working with Swedish Bird Recoveries Data in R
}
\details{
The data included is provided from The Swedish Natural History Museum.
}
\examples{
#see functions.
}
\keyword{package}
