% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries_eng}
\alias{birdrecoveries_eng}
\title{Dataset birdrecoveries_eng}
\format{A data frame
\describe{
  \item{name}{Vernacular}
  \item{sciname}{Scientific Name}
  \item{ringing_sex_code}{Bird Sex Code}
  \item{ringing_sex}{Sex}
  \item{ringing_age_code}{Age Code}
  \item{ringing_age}{Age}
  \item{ringing_date}{Ringing Date}
  \item{ringing_lat}{Ringing Latitude}
  \item{ringing_lon}{Ringing Longitude}
  \item{ringing_country}{Ringing Country}
  \item{ringing_province}{Ringing Province}
  \item{ringing_majorplace}{Ringing Location (Approx)}
  \item{ringing_minorplace}{Ringing Location (Precise)}
  \item{recovery_type}{Recovery Type}
  \item{recovery_date}{Recovery Date}
  \item{recovery_date_accu_code}{Recovery Date Accuracy Code}
  \item{recovery_date_accu}{Recovery Date Accuracy}
  \item{recovery_lat}{Recovery Latitude}
  \item{recovery_lon}{Recovery Longitude}
  \item{recovery_coord_accu}{Recovery Location Accuracy}
  \item{recovery_country}{Recovery Country}
  \item{recovery_province}{Recovery Province}
  \item{recovery_majorplace}{Recovery Location (Approx)}
  \item{recovery_minorplace}{Recovery Location (Precise)}
  \item{distance}{Distance (km)}
  \item{direction}{Direction (degrees)}
  \item{days}{Days Since Ringing}
  \item{hours}{Hours}
  \item{recovery_code}{Recovery Code}
  \item{recovery_details}{Recovery Details}
  \item{source}{Source}
  \item{modified_date}{Registration Date}
  ...
}}
\source{
\url{http://fagel3.nrm.se/fagel/aterfynd/SQLDataExport.csv}
}
\usage{
birdrecoveries_eng
}
\description{
Date: 2017-06-24
}
\keyword{datasets}
