% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries_i18n}
\alias{birdrecoveries_i18n}
\title{Dataset birdrecoveries_i18n}
\format{A data frame
\describe{
  \item{colname}{Field}
  \item{desc_swe}{Translation in Swedish}
  \item{desc_eng}{Translation in English}
  ...
  }}
\source{
\url{}
}
\usage{
birdrecoveries_i18n
}
\description{
Date: 2017-06-24
}
\keyword{datasets}
