% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdata.R
\docType{data}
\name{birdrecoveries_swe}
\alias{birdrecoveries_swe}
\title{Dataset birdrecoveries_swe}
\format{A data frame [166262 x 31]
\describe{
  \item{name}{Art}
  \item{sciname}{Binomen}
  \item{ringing_sex_code}{Könskod}
  \item{ringing_sex}{Kön}
  \item{ringing_age_code}{Ålderskod}
  \item{ringing_age}{Ålder}
  \item{ringing_date}{Märkdatum}
  \item{ringing_lat}{Märklatitud}
  \item{ringing_lon}{Märklongitud}
  \item{ringing_country}{Märkland}
  \item{ringing_province}{Märkprovins}
  \item{ringing_majorregion}{Märkningens Storregion}
  \item{ringing_minorregion}{Märkningens Småregion}
  \item{recovery_type}{Fyndets typ}
  \item{recovery_date}{Fyndets datum}
  \item{recovery_date_accu_code}{Fynddatumets nogrannhetskod}
  \item{recovery_date_accu}{Fynddatumets noggrannhet}
  \item{recovery_lat}{Fyndets latitud}
  \item{recovery_lon}{Fyndets longitud}
  \item{recovery_coord_accu}{Fyndkoordinatens noggrannhet}
  \item{recovery_country}{Fyndland}
  \item{recovery_province}{Fyndprovins}
  \item{recovery_majorregion}{Fyndets storregion}
  \item{recovery_minorregion}{Fyndets småregion}
  \item{distance}{Avstånd}
  \item{direction}{Riktning}
  \item{days}{Dagar sedan märkning}
  \item{hours}{Timmar}
  \item{recovery_code}{Fyndkod}
  \item{recovery_details}{Fynddetaljer}
  \item{source}{Källa}
  ...
  }}
\source{
\url{https://archive.org/download/swedishbirdrecoveries/recoveries.xlsx}
}
\usage{
birdrecoveries_swe
}
\description{
Date: Mon Nov 21 10:16:48 2016
}
\keyword{datasets}

