% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_falsterbo}
\alias{scrape_checklist_falsterbo}
\title{Retrieve species list from Falsterbo Fågelstation}
\usage{
scrape_checklist_falsterbo()
}
\value{
a data frame with species names and codes
}
\description{
Retrieve species list from Falsterbo Fågelstation
}
\examples{
df <- scrape_checklist_falsterbo()
}
