% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_norway}
\alias{scrape_checklist_norway}
\title{Retrieve species list from Norway, Stavanger}
\usage{
scrape_checklist_norway()
}
\value{
a data frame
}
\description{
Retrieve species list from Norway, Stavanger
}
\examples{
df <- scrape_checklist_norway()
}
