% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_checklist_ottenby}
\alias{scrape_checklist_ottenby}
\title{Retrieve species list from Ottenby Fågelstation}
\usage{
scrape_checklist_ottenby()
}
\value{
a data frame with species names and codes
}
\description{
Retrieve species list from Ottenby Fågelstation
}
\examples{
df <- scrape_checklist_ottenby()
}
