% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_recoveries_norway}
\alias{scrape_recoveries_norway}
\title{Retrieve bird recovery data from Norway (Stavanger)}
\usage{
scrape_recoveries_norway(species_id = 1580)
}
\arguments{
\item{species_id}{the species id for a bird species}
}
\value{
a data frame with recovery data
}
\description{
Retrieve bird recovery data from Norway (Stavanger)
}
\examples{
df <- scrape_recoveries_norway()
}
