% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_recoveries_ottenby}
\alias{scrape_recoveries_ottenby}
\title{Retrieve bird recovery data from Ottenby Fågelstation}
\usage{
scrape_recoveries_ottenby()
}
\value{
a data frame with recovery data
}
\description{
Retrieve bird recovery data from Ottenby Fågelstation
}
\examples{
df <- scrape_recoveries_ottenby()
}

