% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_data}
\alias{update_data}
\title{Update local db with dataset from remote, create if needed}
\usage{
update_data(force = FALSE)
}
\description{
Update local db with dataset from remote, create if needed
}
